Fix the bug about resolving the
