+++ src/runners/baseRunner/BaseRunner.ts
@@ -13,8 +13,8 @@ import { ITestItem } from '../../protocols';
 import { IExecutionConfig } from '../../runConfigs';
 import * as classpathUtils from '../../utils/classpathUtils';
 import { resolveRuntimeClassPath } from '../../utils/commandUtils';
-import { resolveWorkingDirectory } from '../../utils/configUtils';
 import { killProcess } from '../../utils/cpUtils';
+import { resolveWorkingDirectory } from '../../utils/settingUtils';
 import { ITestRunner } from '../ITestRunner';
 import { ITestResult } from '../models';
 import { BaseRunnerResultAnalyzer } from './BaseRunnerResultAnalyzer';
